lambda/requirements.txt
boto3==1.34.69
ask-sdk-core==1.19.0
python-dotenv==1.0.1
requests==2.31.0
